requirements.txt
firebase-admin
pyrebase4
google-genai
dotenv
flask
flask-restful
flask-cors
setuptools
flask-socketio
eventlet
celery
redis